feat(backend): roleId 改为 Long 类型

+++ admin-backend/src/main/java/com/example/backend/common/Utils/SessionUtils.java
@@ -60,7 +60,7 @@ public class SessionUtils {
         return userId;
     }
 
-    public static Integer getRoleId(HttpSession session) {
+    public static Long getRoleId(HttpSession session) {
         Integer roleId = getInteger(session, "role_id");
         return roleId;
     }
